requirements.txt
django~=4.0.1
djangorestframework~=3.13.1
django-rest-passwordreset
psycopg2-binary
pyyaml~=6.0
ujson
requestsasgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-cors-headers==3.11.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
idna==3.3
psycopg2-binary==2.9.3
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
ujson==5.1.0
urllib3==1.26.8
celery~=5.2.3